Create DATA-CLEANING-IN-MYSQL
This shows layoffs all around the world. By cleaning this data, I intend to later uncover trends and patterns within the layoffs, which are significant for understanding the broader economic landscape and potentially identifying factors that contribute to layoffs.

Through this project, you hope to gain insights that could inform decision-making processes and perhaps even help mitigate the impact of layoffs in the future.